+++ tests/conftest.py
@@ -6,6 +6,8 @@ class FakeRequest:
 
     def __init__(self, data):
         self._data = data
+        self.base_url = "http://localhost:8080"
+        self.headers = {"Host": "localhost:8080"}
 
     def get_json(self, silent=False):
         return self._data